+++ docs/main/changelog.rst
@@ -12,7 +12,7 @@ Development version
 -------------------
 
 - Add functionality for decorating classes with ``@di.inject``.
-- Add functionality for creating ``di.AbstractCatalog`` subsets.
+- Add functionality for creating ``di.AbstractCatalog`` provider bundles.
 - Add enhancement for ``di.AbstractCatalog`` inheritance.
 - Add images for catalog "Writing catalogs" and "Operating with catalogs" 
   examples.